+++ src/Sulu/Bundle/ContactBundle/SuluContactBundle.php
@@ -16,10 +16,5 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 class SuluContactBundle extends Bundle
 {
-    public function build(ContainerBuilder $container)
-    {
-        parent::build($container);
 
-        $container->addCompilerPass(new AddContentNavigationPass());
-    }
 }