+++ src/Sulu/Bundle/SnippetBundle/Resources/public/js/components/snippet/form/main.js
@@ -46,7 +46,7 @@ define([
             breadcrumb: this.breadcrumb,
 
             tabs: {
-                url: '/admin/snippet/navigation/snippet'
+                url: '/admin/content-navigations?alias=snippet'
             },
 
             toolbar: {